Added nodemailer to server/api/controllers/user.js
I added nodemailer logic. User does not properly get redirected on account creation. Email is sent successfully.

Problem is that I added the entirety of the html in the object before sending it, tried to do it as an import but was struggling with the logic. Added the file separately in an html folder to be used as an import.